Introduce ABC RPC functionality
Summary: As per title.

Test Plan: Not sure. I assume we'd need to add a pythin test for it.

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Differential Revision: https://reviews.bitcoinabc.org/D17

+++ src/globals.cpp
@@ -6,4 +6,4 @@
 
 #include "consensus/consensus.h"
 
-unsigned int nMaxBlockSize = DEFAULT_MAX_BLOCK_SIZE;
+uint64_t nMaxBlockSize = DEFAULT_MAX_BLOCK_SIZE;
