+++ src/globals.h
@@ -5,7 +5,9 @@
 #ifndef BITCOIN_GLOBALS_H
 #define BITCOIN_GLOBALS_H
 
+#include <cstdint>
+
 /** The largest block size this node will accept. */
-extern unsigned int nMaxBlockSize;
+extern uint64_t nMaxBlockSize;
 
 #endif // BITCOIN_GLOBALS_H
